+++ app/src/main/java/luubieunghi/lbn/booklib/Model/BookFile/BookFile.java
@@ -17,7 +17,7 @@ public class BookFile {
 
     //Các biến dùng cho RSL.
     @ColumnInfo(name = "bookID")
-    private int bookID;
+    private long bookID;
 
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "bFileOrder")
@@ -32,11 +32,11 @@ public class BookFile {
     @ColumnInfo(name = "bTotal")
     private long bTotal;
 
-    public int getBookID() {
+    public long getBookID() {
         return bookID;
     }
 
-    public void setBookID(int bookID) {
+    public void setBookID(long bookID) {
         this.bookID = bookID;
     }
